Fix fresh migration failures
This fixes migration failures for a new database by making the
`CanvasCourse.objects.all()` query no longer happen immediatly on app start.

The query happening immediatly would lead to crashes when the model did not
yet exist as django would not get the chance to create the model before the
query was executed.